+++ marshmallow/utils.py
@@ -12,7 +12,6 @@ from calendar import timegm
 from decimal import Decimal, ROUND_HALF_EVEN, Context, Inexact
 from email.utils import formatdate, parsedate
 from pprint import pprint as py_pprint
-import warnings
 
 from marshmallow.compat import OrderedDict, binary_type, text_type
 
@@ -317,7 +316,8 @@ def _get_value_for_key(key, obj, default):
         return obj[key]
     except (KeyError, AttributeError, IndexError, TypeError):
         try:
-            return getattr(obj, key)
+            attr = getattr(obj, key)
+            return attr() if callable(attr) else attr
         except AttributeError:
             return default
     return default
@@ -333,17 +333,6 @@ def get_callable_name(func):
         return func.func.__name__
     else:  # Callable class
         return func.__class__.__name__
-
-
-def get_func_name(func):
-    """Given a callable, return its name. Handles `functools.partial` objects.
-
-    .. deprecated:: 1.2.0
-        Renamed to :func:`get_callable_name`.
-    """
-    warnings.warn('get_func_name was renamed to get_callable_name. get_func_name '
-                  'will be removed in version 2.0', DeprecationWarning)
-    return get_callable_name(func)
 
 
 def callable_or_raise(obj):